Split config cli endpoint to package
