+++ src/statue/cli/config/__init__.py
@@ -1,0 +1,12 @@
+"""Configuration related CLIs."""
+from statue.cli.config.config_cli import config_cli
+from statue.cli.config.config_commands import fixate_commands_versions_cli
+from statue.cli.config.config_init import init_config_cli
+from statue.cli.config.config_show import show_config_tree_cli
+
+__all__ = [
+    "config_cli",
+    "fixate_commands_versions_cli",
+    "init_config_cli",
+    "show_config_tree_cli",
+]
