+++ src/statue/cli/config/config_cli.py
@@ -1,0 +1,9 @@
+"""Configuration main CLI."""
+
+
+from statue.cli.cli import statue_cli
+
+
+@statue_cli.group("config")
+def config_cli():
+    """Configuration related actions."""
